Fix bug where scoreboard labels would change to incorrect player names
Source of the problem:
- A closure has references to the variable in its enclosing scope. In this
case, "bird".  Now, the value of "bird" gets mutated after the closures
are created simply as a part of basic iteration. So, all of the closures
reference the same variable, "bird". At the time the closures execute, the
for loop is complete and bird has the value it had on the final iteration
through the for loop.

So, the fix was to create a local variable in the enclosing scope so that on
each iteration through the loop, each closure is referencing a different variable.